requirements.txt
pandas==1.3.0
numpy>=1.18.5
pytest==7.3.1
torch==1.13.1
scikit-learn==0.22.2
deprecation==2.1.0
fasttext-wheel==0.9.2
gensim==3.8.3
Cython==0.29.14


